fix: woocommerce shop description markup
WooCommerce adds some extra markup around any content added through the editor on the shop page. This extra markup, and accompanying CSS, messes with the theme's built-in styling and can result in normal paragraphs having a huge font size, bold and other unpleasantness.

This change overrides the relevant Woo function to echo the content as-is without any extra markup.